requirements.txt
mysql-connector-python
tk
pyinstaller